Mechinmind/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-markdownx==3.0.1
Markdown==3.3.4
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
